+++ app/src/main/java/com/dessalines/rankmyfavs/utils/Utils.kt
@@ -1,5 +1,4 @@
 package com.dessalines.rankmyfavs.utils
-
 import android.content.ContentResolver
 import android.content.Context
 import android.content.Intent
@@ -8,6 +7,8 @@ import android.content.pm.PackageManager
 import android.graphics.Bitmap
 import android.net.Uri
 import android.os.Build
+import androidx.compose.ui.graphics.Color
+import com.dessalines.rankmyfavs.db.FavListItem
 import java.io.IOException
 import java.io.OutputStream
 
@@ -21,6 +22,15 @@ const val MASTODON_URL = "https://mastodon.social/@dessalines"
 const val GLICKO_WIKI_URL = "https://en.m.wikipedia.org/wiki/Glicko_rating_system"
 
 const val ANIMATION_SPEED = 300
+
+val TIER_COLORS =
+    mapOf(
+        "S" to Color(0XFFFF7F7F),
+        "A" to Color(0XFFFFBF7F),
+        "B" to Color(0XFFFFDF7F),
+        "C" to Color(0XFFFFFF7F),
+        "D" to Color(0XFFBFFF7F),
+    )
 
 fun openLink(
     url: String,
@@ -51,11 +61,11 @@ fun numToString(
 ): String = String.format("%.${decimalPlaces}f", num)
 
 fun writeData(
-    contentResolver: ContentResolver,
+    ctx: Context,
     uri: Uri,
     data: String,
 ) {
-    contentResolver.openOutputStream(uri)?.use {
+    ctx.contentResolver.openOutputStream(uri)?.use {
         val bytes = data.toByteArray()
         it.write(bytes)
     }
@@ -74,4 +84,14 @@ fun writeBitmap(
     } catch (e: IOException) {
         e.printStackTrace()
     }
+}
+
+fun nameIsValid(name: String): Boolean = name.isNotEmpty()
+
+fun convertFavlistToMarkdown(
+    title: String,
+    favListItems: List<FavListItem>,
+): String {
+    val items = favListItems.joinToString(separator = "\n") { "1. ${it.name}" }
+    return "# $title\n\n$items"
 }